# docs/swagger.yaml
basePath: /
definitions:
  controllers.authResponse:
    properties:
      message:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/controllers.userResponse'
    type: object
...
  controllers.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  controllers.registerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  controllers.userResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  httpapi.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  httpapi.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/httpapi.ErrorDetail'
    type: object
info:
  contact: {}
  description: Personal Finance Tracker API with JWT auth, pagination, filtering,
    and observability hooks.
  title: Personal Finance Tracker API
  version: "1.0"
paths:
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a signed auth token.
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/controllers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Log in a user
      tags:
      - auth
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Create a user account and return a signed auth token.
      parameters:
      - description: Registration payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/controllers.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Provide the JWT as `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"